requirements.txt
Flask
Pillow
gunicorn
google-generativeai
python-dotenv
